Refactor the list and build docs
This makes some pretty significant changes to the way the list works to
just generally be better. Rather than returning `bool` for
success/failure, instead zero is now returned on success and non-zero is
returned on failure for most functions (except for those where
true/false really does make sense).

`list_destroy()` now takes a second argument which is a function that
knows how to destroy the items within the list. This makes it safer to
call `list_destroy()` directly when list items have been heap-allocated.

Documentation is now generated with Doxygen for all functions and types
in the header file.

Reversing the list is now directly supported as a function in the header
file.